+++ request/utils.py
@@ -160,7 +160,12 @@ def get_verbose_name(class_name):
 def request_is_ajax(request):
     return (
         request.META.get('HTTP_X_REQUESTED_WITH') == 'XMLHttpRequest' or
-        request.META.get('HTTP_HX_REQUEST') == 'true'  # htmx
+        # Htmx.
+        (
+            request.META.get('HTTP_HX_REQUEST') == 'true' and
+            # Do not ignore boosted htmx requests.
+            request.META.get('HTTP_HX_BOOSTED') != 'true'
+        )
     )